structure almost ready
