--- src/main/scala/art_of_joy/domain/PersonLayer.scala
@@ -1,225 +1,0 @@
-package art_of_joy.domain
-
-import zio.{Scope, ZIO, ZLayer}
-import io.getquill.*
-
-import javax.sql.DataSource
-import art_of_joy.ctx
-import art_of_joy.model.`enum`.{RegistrationError, Role, SetPasswordError}
-import art_of_joy.model.http.{HttpResponse, HttpValidationFields}
-import art_of_joy.model.person.{SetPassword, Person, RegPerson}
-import art_of_joy.domain.SessionStorageLayer.StoragePerson
-import art_of_joy.domain.service.interfaces.{EmailService, PersonService, SessionStorageService}
-import art_of_joy.utils.*
-
-import java.util.{Date, UUID}
-object PersonLayer {
-  import ctx._
-  private def addPerson(email:String, password:String, phone:String): ZIO[DataSource, Throwable, Person] =
-    for {
-      user <- ctx.run(
-        quote {
-          query[Person].insert(_.phone -> lift(Option(phone)),
-            _.role -> lift(Role.user.ordinal),
-            _.email -> lift(Option(email)),
-            _.password_hash -> lift(Option(password)),
-            _.is_confirm_email -> false,
-            _.is_confirm_phone -> false
-          ).returning(p => p)
-        }
-      )
-    } yield user
-
-  private def getRegistrationError(passwordValid:Boolean, emailValid:Boolean, emailChecker:Boolean,phoneChecker:Boolean) =
-    List(
-      (passwordValid,"password_regFormTI", RegistrationError.passwordValidationError),
-      (emailValid,"email_regFormTI", RegistrationError.emailValidationError),
-      (emailChecker,"email_regFormTI", RegistrationError.emailCheckerError),
-      (phoneChecker,"phone_regFormTI", RegistrationError.phoneCheckerError)
-    ).collect{ case (false, fieldName, msg) => HttpValidationFields(fieldName, msg.message)}
-
-  private def getPasswordError(repeatValid:Boolean, passwordValid:Boolean, isOldEqual:Boolean, oldEqualNewValid:Boolean) =
-    List(
-      (repeatValid, "passwordRepeat_userInfoFormTI", "Пароли не совпадают"),
-      (passwordValid, "password_userInfoFormTI", RegistrationError.passwordValidationError.message),
-      (isOldEqual, "oldPassword_userInfoFormTI", "Неверный старый пароль"),
-      (oldEqualNewValid, "oldPassword_userInfoFormTI", "Новый пароль совпадает со старым")
-    ).collect{ case (false, fieldName, msg) => HttpValidationFields(fieldName, msg)}
-
-  private def getPassword(setPassword: SetPassword):List[HttpValidationFields] = ???
-
-  val live = ZLayer.succeed(
-    new PersonService {
-      override def getAllPersons(startRow:Int, endRow:Option[Int]): ZIO[DataSource, Throwable, List[Person]] =
-        for{
-          users <- endRow match
-            case Some(value) =>
-              for{
-                _ <- ZIO.when(value < startRow)(ZIO.fail(new Exception("endRow должен быть больше startRow")))
-                users <- ctx.run(
-                  quote{
-                    query[Person].drop(lift(startRow)).take(lift(value))
-                  }
-                )
-              }yield users
-            case None => ctx.run(
-              quote{
-                query[Person]
-              }
-            )
-        }yield users
-
-      override def getPersonByEmail(email:String): ZIO[DataSource, Throwable, List[Person]] =
-        ctx.run(
-          quote{
-            query[Person].filter(p =>
-              p.email == lift(Option(email))
-            )
-          }
-        )
-
-      override def authPerson(email: String, password: String): ZIO[DataSource, Throwable, List[Person]] =
-        ctx.run(
-          quote{
-            query[Person].filter(p => p.email == lift(Option(email)) && p.password_hash.getOrElse("") == lift(passToHash(password)))
-          }
-        )
-
-      override def checkEmail(email: String): ZIO[DataSource, Throwable, Boolean] =
-        ctx.run(
-          quote{
-            query[Person].filter(p => p.email == lift(Option(email)) )
-          }
-        ).map(_.isEmpty)
-
-
-      override def checkPhone(phone: String): ZIO[DataSource, Throwable, Boolean] =
-        ctx.run(
-          quote {
-            query[Person].filter(p => p.phone == lift(Option(phone)))
-          }
-        ).map(_.isEmpty)
-
-      override def emailRegistration(email: String): ZIO[SessionStorageService & DataSource & EmailService, Throwable, Either[String, List[HttpValidationFields]]] =
-        for{
-          emailValid <- ZIO.from(isValidEmail(email))
-          isEmailBusy <- checkEmail(email)
-          result <- ZIO.ifZIO(ZIO.from(!isEmailBusy))(
-            onTrue =ZIO.from(
-              getRegistrationError(true, emailValid, isEmailBusy, true)
-            ).map(Right(_)),
-            onFalse =
-              (
-                for{
-                  storage <- ZIO.service[SessionStorageService]
-                  sessionID <- ZIO.from(UUID.randomUUID.toString)
-                  acceptCode <- ZIO.succeed(generateCode)
-                  _ <- EmailService.sendMessage("Подтверждение почты", s"Ваш код подтверждения $acceptCode", email)
-                  _ <- storage.put(sessionID, StoragePerson(
-                    Person(id = -1, email = Option(email), role = Role.user.ordinal, is_confirm_email = false, is_confirm_phone = false),
-                    new Date().getTime, acceptCode
-                  ))
-                }yield sessionID
-              ).map(Left(_))
-          )
-        }yield result
-
-      override def phoneRegistration(phone: String): ZIO[SessionStorageService, Throwable, Either[String, List[HttpValidationFields]]] = ???
-
-      override def addPerson(person: Person): ZIO[DataSource, Throwable, Person] =
-        ctx.run(
-          quote {
-            query[Person].insert(
-              _.role -> lift(person.role),
-              _.phone -> lift(person.phone),
-              _.email -> lift(person.email),
-              _.surname -> lift(person.surname),
-              _.firstname -> lift(person.firstname),
-              _.middlename -> lift(person.middlename),
-              _.is_confirm_email -> lift(person.is_confirm_email),
-              _.is_confirm_phone -> lift(person.is_confirm_phone),
-            ).returning(p => p)
-          }
-        )
-
-      override def authPersonOnEmail(email: String): ZIO[DataSource & SessionStorageService, Throwable, String] =
-        for{
-          isRegistration <- checkEmail(email)
-          _ <- ZIO.when(isRegistration)(ZIO.fail(new Exception("Пользователь с такой почтой не зарегистрирован")))
-          service <- ZIO.service[SessionStorageService]
-          token <- ZIO.from(
-            UUID.randomUUID.toString
-          )
-          _ <- service.put(token, StoragePerson(
-            Person(email = Option(email)), new Date().getTime, generateCode
-          ))
-        }yield token
-
-      override def setPassword(personID:Int, newPassword:String, repeatPassword:String):ZIO[DataSource, Throwable, Either[Long, List[HttpValidationFields]]] =
-        for{
-          validationError <- ZIO.from(
-            getPasswordError(
-              newPassword == repeatPassword,
-              isValidPassword(newPassword),
-              true,
-              true
-            )
-          )
-          result <-
-            if (validationError.isEmpty)
-                ctx.run(
-                  quote{
-                    query[Person].filter(_.id == lift(personID)).update(_.password_hash -> lift(Option(passToHash(newPassword))))
-                  }
-                ).map(Left(_))
-            else
-              ZIO.from(validationError).map(Right(_))
-        }yield result
-
-      override def getPersonByID(personID: Long): ZIO[DataSource, Throwable, Person] =
-        ctx.run(
-          quote{
-            query[Person].filter(_.id == lift(personID))
-          }
-        ).map(_.head)
-        
-      override def updatePassword(personID: Int, newPassword: String, repeatPassword: String, oldPassword: String): ZIO[DataSource, Throwable, Either[Long, List[HttpValidationFields]]] =
-        for{
-          dbPersonPass <- getPersonByID(personID)
-          newPassHash <- ZIO.from(passToHash(newPassword))
-          validationError <- ZIO.from(
-            getPasswordError(
-              newPassword == repeatPassword,
-              isValidPassword(newPassword),
-              newPassHash == dbPersonPass.password_hash.get,
-              newPassword != oldPassword
-            )
-          )
-          result <-
-            if (validationError.isEmpty)
-              ctx.run(
-                quote{
-                  query[Person].filter(_.id == lift(personID)).update(_.password_hash -> lift(Option(newPassHash)))
-                }
-              ).map(Left(_))
-            else
-              ZIO.from(validationError).map(Right(_))
-        }yield result
-
-      override def setPersonInfo(id:Long, surname:String, firstname:String, middleName:Option[String]): ZIO[DataSource, Throwable, Long] =
-        ctx.run(
-          dynamicQuery[Person]
-            .filter(_.id == lift(id))
-            .update(
-              setValue(_.firstname, Option(firstname)),
-              setValue(_.surname, Option(surname)),
-              setOpt(_.middlename, middleName match
-                case Some(value) => Option(Option(value))
-                case None => Option.empty[Option[String]]
-              )
-            )
-        )
-
-    }
-  )
-}
