+++ src/main/scala/art_of_joy/domain/service/email/Email.scala
@@ -1,0 +1,11 @@
+package art_of_joy.domain.service.email
+
+import zio._
+
+trait Email {
+  def sendMessage(title:String, body:String,  emailReceiver:String):ZIO[Any,Throwable, Unit]
+}
+object Email{
+  def sendMessage(title:String, body:String,  emailReceiver:String) =
+    ZIO.serviceWithZIO[Email](_.sendMessage(title,body,emailReceiver))
+}
